Validation and other stuff

+++ src/main/java/by/triumgroup/recourse/controller/impl/AbstractCrudController.java
@@ -1,6 +1,9 @@
 package by.triumgroup.recourse.controller.impl;
 
+import by.triumgroup.recourse.configuration.security.Auth;
+import by.triumgroup.recourse.configuration.security.UserAuthDetails;
 import by.triumgroup.recourse.controller.CrudController;
+import by.triumgroup.recourse.controller.exception.AccessDeniedException;
 import by.triumgroup.recourse.controller.exception.BadRequestException;
 import by.triumgroup.recourse.controller.exception.NotFoundException;
 import by.triumgroup.recourse.entity.model.BaseEntity;
@@ -11,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 
 import javax.validation.Valid;
 import java.util.Optional;
+import java.util.function.BiFunction;
 
 import static by.triumgroup.recourse.util.ServiceCallWrapper.wrapServiceCall;
 
@@ -24,10 +28,28 @@ public abstract class AbstractCrudController<E extends BaseEntity<ID>, ID> imple
         this.crudService = crudService;
     }
 
-    public E getById(@PathVariable("id") ID id) {
+    protected abstract boolean hasAuthorityToEdit(E entity, UserAuthDetails authDetails);
+
+    protected boolean hasAuthorityToRead(E entity, @Auth UserAuthDetails authDetails) {
+        return true;
+    }
+
+    protected void checkAuthority(E entity, UserAuthDetails authDetails, BiFunction<E, UserAuthDetails, Boolean> authorityChecker) {
+        if (!authDetails.isAdmin() && !authorityChecker.apply(entity, authDetails)) {
+            throw new AccessDeniedException();
+        }
+    }
+
+    @Override
+    public E getById(@PathVariable("id") ID id, @Auth UserAuthDetails userAuthDetails) {
         return wrapServiceCall(logger, () -> {
             Optional<E> callResult = crudService.findById(id);
-            return callResult.orElseThrow(NotFoundException::new);
+            if (callResult.isPresent()){
+                checkAuthority(callResult.get(), userAuthDetails, this::hasAuthorityToRead);
+                return callResult.get();
+            } else {
+                throw new NotFoundException();
+            }
         });
     }
 
@@ -36,7 +58,8 @@ public abstract class AbstractCrudController<E extends BaseEntity<ID>, ID> imple
     }
 
     @Override
-    public <S extends E> S create(@Valid @RequestBody S entity) {
+    public <S extends E> S create(@Valid @RequestBody S entity, @Auth UserAuthDetails authDetails) {
+        checkAuthority(entity, authDetails, this::hasAuthorityToEdit);
         return wrapServiceCall(logger, () -> {
             Optional<S> callResult = crudService.add(entity);
             return callResult.orElseThrow(BadRequestException::new);
@@ -44,7 +67,8 @@ public abstract class AbstractCrudController<E extends BaseEntity<ID>, ID> imple
     }
 
     @Override
-    public <S extends E> S update(@Valid @RequestBody S entity, @PathVariable("id") ID id) {
+    public <S extends E> S update(@Valid @RequestBody S entity, @PathVariable("id") ID id, @Auth UserAuthDetails authDetails) {
+        checkAuthority(entity, authDetails, this::hasAuthorityToEdit);
         return wrapServiceCall(logger, () -> {
             Optional<S> callResult = crudService.update(entity, id);
             return callResult.orElseThrow(NotFoundException::new);
@@ -52,7 +76,8 @@ public abstract class AbstractCrudController<E extends BaseEntity<ID>, ID> imple
     }
 
     @Override
-    public void delete(@PathVariable("id") ID id) {
+    public void delete(@PathVariable("id") ID id, @Auth UserAuthDetails authDetails) {
+        checkAuthority(getById(id, authDetails), authDetails, this::hasAuthorityToEdit);
         wrapServiceCall(logger, () -> crudService.delete(id).orElseThrow(NotFoundException::new));
     }
 }
